pyproject.toml
[project]
name = "strands-agent-lambda"
version = "1.0.0"
description = "Strands Agent on AWS Lambda"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "strands-agents>=0.1.7",
    "strands-agents-tools>=0.1.5",
]

[project.optional-dependencies]
dev = [
    "aws-cdk-lib>=2.100.0",
    "constructs>=10.0.0",
]

[tool.uv]
dev-dependencies = [
    "aws-cdk-lib>=2.100.0",
    "constructs>=10.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["lambda"]